Break up fold.cc along per-type fault lines
